Added partial 'no-connection' support

+++ lib/data/repository/notes_repository_impl.dart
@@ -1,4 +1,6 @@
 import 'package:f_logs/f_logs.dart';
+import 'package:registro_elettronico/core/error/failures.dart';
+import 'package:registro_elettronico/core/network/network_info.dart';
 import 'package:registro_elettronico/data/db/dao/note_dao.dart';
 import 'package:registro_elettronico/data/db/dao/profile_dao.dart';
 import 'package:registro_elettronico/data/db/moor_database.dart';
@@ -11,10 +13,13 @@ class NotesRepositoryImpl implements NotesRepository {
   NoteDao noteDao;
   SpaggiariClient spaggiariClient;
   ProfileDao profileDao;
+  NetworkInfo networkInfo;
+
   NotesRepositoryImpl(
     this.noteDao,
     this.spaggiariClient,
     this.profileDao,
+    this.networkInfo,
   );
 
   @override
@@ -34,48 +39,57 @@ class NotesRepositoryImpl implements NotesRepository {
 
   @override
   Future updateNotes() async {
-    final profile = await profileDao.getProfile();
-    await noteDao.deleteAllAttachments();
-    await noteDao.deleteAllNotes();
-    final notesResponse = await spaggiariClient.getNotes(profile.studentId);
+    if (await networkInfo.isConnected) {
+      final profile = await profileDao.getProfile();
 
-    List<Note> notes = [];
-    notesResponse.notesNTCL.forEach((note) =>
-        notes.add(NoteMapper.convertNotetEntityToInsertable(note, 'NTCL')));
-    notesResponse.notesNTWN.forEach((note) =>
-        notes.add(NoteMapper.convertNotetEntityToInsertable(note, 'NTWN')));
-    notesResponse.notesNTTE.forEach((note) =>
-        notes.add(NoteMapper.convertNotetEntityToInsertable(note, 'NTTE')));
-    notesResponse.notesNTST.forEach((note) =>
-        notes.add(NoteMapper.convertNotetEntityToInsertable(note, 'NTST')));
+      final notesResponse = await spaggiariClient.getNotes(profile.studentId);
 
-    FLog.info(
-      text:
-          'Got ${notesResponse.notesNTCL.length} notesNTCL from server, procceding to insert in database',
-    );
-    FLog.info(
-      text:
-          'Got ${notesResponse.notesNTWN.length} notesNTWN from server, procceding to insert in database',
-    );
-    FLog.info(
-      text:
-          'Got ${notesResponse.notesNTTE.length} notesNTTE from server, procceding to insert in database',
-    );
-    FLog.info(
-      text:
-          'Got ${notesResponse.notesNTST.length} notesNTST from server, procceding to insert in database',
-    );
+      List<Note> notes = [];
+      notesResponse.notesNTCL.forEach((note) =>
+          notes.add(NoteMapper.convertNotetEntityToInsertable(note, 'NTCL')));
+      notesResponse.notesNTWN.forEach((note) =>
+          notes.add(NoteMapper.convertNotetEntityToInsertable(note, 'NTWN')));
+      notesResponse.notesNTTE.forEach((note) =>
+          notes.add(NoteMapper.convertNotetEntityToInsertable(note, 'NTTE')));
+      notesResponse.notesNTST.forEach((note) =>
+          notes.add(NoteMapper.convertNotetEntityToInsertable(note, 'NTST')));
 
-    await noteDao.insertNotes(notes);
+      FLog.info(
+        text:
+            'Got ${notesResponse.notesNTCL.length} notesNTCL from server, procceding to insert in database',
+      );
+      FLog.info(
+        text:
+            'Got ${notesResponse.notesNTWN.length} notesNTWN from server, procceding to insert in database',
+      );
+      FLog.info(
+        text:
+            'Got ${notesResponse.notesNTTE.length} notesNTTE from server, procceding to insert in database',
+      );
+      FLog.info(
+        text:
+            'Got ${notesResponse.notesNTST.length} notesNTST from server, procceding to insert in database',
+      );
+      await noteDao.deleteAllAttachments();
+      await noteDao.deleteAllNotes();
+
+      await noteDao.insertNotes(notes);
+    } else {
+      throw new NotConntectedException();
+    }
   }
 
   @override
   Future<NotesReadResponse> readNote(String type, int eventId) async {
-    final profile = await profileDao.getProfile();
+    if (await networkInfo.isConnected) {
+      final profile = await profileDao.getProfile();
 
-    final res =
-        await spaggiariClient.markNote(profile.studentId, type, eventId, "");
-    return res;
+      final res =
+          await spaggiariClient.markNote(profile.studentId, type, eventId, "");
+      return res;
+    } else {
+      throw new NotConntectedException();
+    }
   }
 
   @override
@@ -85,20 +99,25 @@ class NotesRepositoryImpl implements NotesRepository {
 
   @override
   Future<NotesAttachment> getAttachmentForNote(String type, int eventId) async {
-    final profile = await profileDao.getProfile();
-    final attachments = await noteDao.getAllAttachments();
+    if (await networkInfo.isConnected) {
+      final profile = await profileDao.getProfile();
+      final attachments = await noteDao.getAllAttachments();
 
-    for (var attachment in attachments) {
-      if (attachment.id == eventId) return attachment;
+      for (var attachment in attachments) {
+        if (attachment.id == eventId) return attachment;
+      }
+
+      final res =
+          await spaggiariClient.markNote(profile.studentId, type, eventId, "");
+      final insertable =
+          NoteMapper.convertNoteAttachmentResponseToInsertable(res);
+          
+      await noteDao.deleteAllAttachments();
+      noteDao.insertAttachment(insertable);
+
+      return insertable;
+    } else {
+      throw new NotConntectedException();
     }
-
-    final res =
-        await spaggiariClient.markNote(profile.studentId, type, eventId, "");
-    final insertable =
-        NoteMapper.convertNoteAttachmentResponseToInsertable(res);
-
-    noteDao.insertAttachment(insertable);
-
-    return insertable;
   }
 }
